Nginx configuration now optional.

+++ nginx_profile.sh
@@ -1,4 +1,4 @@
-#!/bin/sh
+#!/bin/bash
 # Copyright Bruno Rahle 2014
 
 NGINX_FOLDER="/etc/nginx"
@@ -15,6 +15,9 @@ sudo sed -ie "s/<SERVER_NAME>/$SERVER_NAME/g" $NGINX_CONF
 sudo sed -ie "s|<STATIC_ROOT>|$VIRTUALENV_DIR|g" $NGINX_CONF
 sudo sed -ie "s/<PORT>/$PROJ_LOCAL_PORT/g" $NGINX_CONF
 
-sudo ln -s -f $NGINX_CONF $SITES_ENABLED/$PROJ_NAME
-sudo service nginx restart
+if confirm "Do you want to add the site to sites-enabled [Y/n] ?"
+then
+    sudo ln -s -f $NGINX_CONF $SITES_ENABLED/$PROJ_NAME
+    sudo service nginx restart
+fi
